Added support for user payload data in push-notification and multiple push-notifications on Android
The implementation only supported id, title, body and silent items, effectively washing out the users payload.  The payload in its entirety is now forwarded and is obtainable from the  Gluon application.  The  hardcoded REQUEST_CODE is now replaced with a unique number to allow for visualizing multiple push-notifications on the Android device.
Both these tasks were handled correctly on the iOS implementation.